move functions out
